Combine filter menu and list layouts into a single one with states
This is preferrable over having different layouts(and hence dialogs) for
them because we don't have much control over dialog backstacks. This was
going to make the code for back button in the filter list dialog ugly.

Using ConstraintLayoutStates allows the potential for using animations
and speeds things up(since a FragmentTransaction is not performed).